Fixed return types on system getStick functions

+++ ttyd-tools/rel/include/ttyd/system.h
@@ -26,9 +26,11 @@ void padRumbleHardOff(uint32_t padId);
 void padRumbleOff(uint32_t padId);
 void padRumbleOn(uint32_t padId);
 
-int8_t keyGetSubStickY(uint32_t padId);
-int8_t keyGetStickY(uint32_t padId);
-int8_t keyGetStickX(uint32_t padId);
+// Signed bytes, but actually returned as uints.
+uint32_t keyGetSubStickY(uint32_t padId);
+uint32_t keyGetStickY(uint32_t padId);
+uint32_t keyGetStickX(uint32_t padId);
+
 uint32_t keyGetButtonTrg(uint32_t padId);
 uint32_t keyGetDirTrg(uint32_t padId);
 uint32_t keyGetButtonRep(uint32_t padId);
